#147 delete unnecessary fragment
