+++ app/src/main/res/layout/activity_follower.xml
@@ -2,11 +2,6 @@
 <layout xmlns:tools="http://schemas.android.com/tools"
     xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto">
-    <data>
-        <variable
-            name="view"
-            type="com.jjbaksa.jjbaksa.ui.follower.FollowerActivity" />
-    </data>
 
     <LinearLayout
         android:layout_width="match_parent"
@@ -38,6 +33,7 @@
                     android:layout_height="48dp"
                     android:layout_marginHorizontal="16dp"
                     app:cardCornerRadius="100dp"
+                    android:layout_marginBottom="24dp"
                     app:layout_constraintTop_toTopOf="parent">
 
                     <androidx.constraintlayout.widget.ConstraintLayout
@@ -74,11 +70,32 @@
                     </androidx.constraintlayout.widget.ConstraintLayout>
                 </androidx.cardview.widget.CardView>
 
-                <androidx.fragment.app.FragmentContainerView
-                    android:id="@+id/fragment_container"
+                <androidx.recyclerview.widget.RecyclerView
+                    android:id="@+id/rv_search_result"
                     android:layout_width="match_parent"
-                    android:layout_height="match_parent" />
+                    android:layout_height="wrap_content"
+                    android:visibility="gone" />
 
+                <androidx.recyclerview.widget.RecyclerView
+                    android:id="@+id/rv_all_follower"
+                    android:layout_width="match_parent"
+                    android:layout_height="wrap_content"
+                    tools:itemCount="4"
+                    tools:listitem="@layout/item_follow" />
+
+                <androidx.recyclerview.widget.RecyclerView
+                    android:id="@+id/rv_request_follow"
+                    android:layout_width="match_parent"
+                    android:layout_height="wrap_content"
+                    android:layout_marginTop="20dp"
+                    tools:itemCount="4"
+                    tools:listitem="@layout/item_follow" />
+
+                <com.jjbaksa.jjbaksa.view.JjLoadingView
+                    android:id="@+id/loading_view"
+                    android:layout_width="match_parent"
+                    android:layout_height="match_parent"
+                    app:loading="true"/>
             </LinearLayout>
 
         </androidx.core.widget.NestedScrollView>
