--- app/src/main/res/layout/fragment_follower.xml
@@ -1,69 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<layout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools">
-
-    <androidx.constraintlayout.widget.ConstraintLayout
-        android:layout_width="match_parent"
-        android:layout_height="match_parent">
-
-
-        <androidx.recyclerview.widget.RecyclerView
-            android:id="@+id/follower_recycler_view"
-            android:layout_width="match_parent"
-            android:layout_height="300dp"
-            android:layout_marginTop="64dp"
-            android:visibility="visible"
-            app:layout_constraintTop_toTopOf="parent"
-            tools:listitem="@layout/item_follow" />
-
-        <View
-            android:id="@+id/view"
-            android:layout_width="match_parent"
-            android:layout_height="1dp"
-            android:background="@color/black"
-            app:layout_constraintBottom_toBottomOf="@+id/follower_recycler_view" />
-
-
-        <androidx.recyclerview.widget.RecyclerView
-            android:id="@+id/recived_follow_request_recycler_view"
-            android:layout_width="match_parent"
-            android:layout_height="0dp"
-            android:layout_marginTop="20dp"
-            android:visibility="visible"
-            app:layout_constraintBottom_toBottomOf="parent"
-            app:layout_constraintTop_toBottomOf="@+id/follower_recycler_view"
-            app:layout_constraintVertical_bias="1.0"
-            tools:layout_editor_absoluteX="0dp"
-            tools:listitem="@layout/item_follow" />
-
-        <androidx.recyclerview.widget.RecyclerView
-            android:id="@+id/send_follow_request_recycler_view"
-            android:layout_width="match_parent"
-            android:layout_height="400dp"
-            android:visibility="visible"
-            app:layout_constraintBottom_toBottomOf="parent"
-            app:layout_constraintTop_toBottomOf="@+id/view"
-            app:layout_constraintVertical_bias="1.0"
-            tools:layout_editor_absoluteX="0dp"
-            tools:listitem="@layout/item_follow" />
-
-
-        <androidx.recyclerview.widget.RecyclerView
-            android:id="@+id/all_users_recycler_view"
-            android:layout_width="match_parent"
-            android:layout_height="0dp"
-            android:visibility="invisible"
-            app:layout_constraintBottom_toBottomOf="parent"
-            app:layout_constraintTop_toTopOf="@+id/follower_recycler_view"
-            tools:listitem="@layout/item_follow" />
-
-        <com.jjbaksa.jjbaksa.view.JjLoadingView
-            android:id="@+id/loading_view"
-            android:layout_width="match_parent"
-            android:layout_height="match_parent"
-            app:loading="true"
-            tools:layout_editor_absoluteX="234dp"
-            tools:layout_editor_absoluteY="181dp" />
-    </androidx.constraintlayout.widget.ConstraintLayout>
-</layout>
